Extracting Multi module
